requirements.txt
Django==5.2.10
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
inflection==0.5.1
mssql-django==1.6
PyJWT==2.10.1
pyodbc==5.3.0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0